auth-sso/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
pydantic==2.7.1
